fix. 이벤트 버블링 이슈 해결 및 기타 디자인 변경
