+++ package.json
@@ -24,6 +24,7 @@
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.51.3",
+    "react-use": "^17.5.0",
     "zustand": "^4.5.2"
   },
   "devDependencies": {
